Adding a feature that allow to run an influxql script on the first docker run

+++ 0.9/run.sh
@@ -98,46 +98,60 @@ else
 fi
 
 # Pre create database on the initiation of the container
-if [ -n "${PRE_CREATE_DB}" ]; then
+if [ -n "${PRE_CREATE_DB}" ] && [ ! -f "/data/.init_script_executed" ]; then
     echo "=> About to create the following database: ${PRE_CREATE_DB}"
-    if [ -f "/data/.pre_db_created" ]; then
-        echo "=> Database had been created before, skipping ..."
+    arr=$(echo ${PRE_CREATE_DB} | tr ";" "\n")
+
+    if [ -n "${ADMIN_USER}" ]; then
+      for x in $arr
+      do
+          echo "=> Creating database: ${x}"
+          echo "CREATE DATABASE ${x}" >> /tmp/init_script.influxql
+          echo "GRANT ALL PRIVILEGES ON ${x} TO ${ADMIN_USER}" >> /tmp/init_script.influxql
+      done
     else
-        arr=$(echo ${PRE_CREATE_DB} | tr ";" "\n")
-
-        #wait for the startup of influxdb
-        RET=1
-        while [[ RET -ne 0 ]]; do
-            echo "=> Waiting for confirmation of InfluxDB service startup ..."
-            sleep 3
-            curl -k ${API_URL}/ping 2> /dev/null
-            RET=$?
-        done
-        echo ""
-
-        PASS=${INFLUXDB_INIT_PWD:-root}
-        if [ -n "${ADMIN_USER}" ]; then
-          echo "=> Creating admin user"
-          influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -execute="CREATE USER ${ADMIN_USER} WITH PASSWORD '${PASS}' WITH ALL PRIVILEGES"
-          for x in $arr
-          do
-              echo "=> Creating database: ${x}"
-              influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -username=${ADMIN_USER} -password="${PASS}" -execute="create database ${x}"
-              influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -username=${ADMIN_USER} -password="${PASS}" -execute="grant all PRIVILEGES on ${x} to ${ADMIN_USER}"
-          done
-          echo ""
-        else
-          for x in $arr
-          do
-              echo "=> Creating database: ${x}"
-              influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -execute="create database \"${x}\""
-          done
-        fi
-
-        touch "/data/.pre_db_created"
+      for x in $arr
+      do
+          echo "=> Creating database: ${x}"
+          echo "CREATE DATABASE ${x}" >> /tmp/init_script.influxql
+      done
     fi
+fi
+
+# Execute influxql queries contained inside /init_script.influxql
+if [ -f "/data/.init_script_executed" ]; then
+    echo "=> The initialization script had been executed before, skipping ..."
+elif [ -f "/init_script.influxql" ] || [ -f "/tmp/init_script.influxql" ]; then
+    echo "=> About to execute the initialization script"
+
+    cat /init_script.influxql >> /tmp/init_script.influxql
+
+    #wait for the startup of influxdb
+    RET=1
+    while [[ RET -ne 0 ]]; do
+        echo "=> Waiting for confirmation of InfluxDB service startup ..."
+        sleep 3
+        curl -k ${API_URL}/ping 2> /dev/null
+        RET=$?
+    done
+    echo ""
+
+    PASS=${INFLUXDB_INIT_PWD:-root}
+    if [ -n "${ADMIN_USER}" ]; then
+      echo "=> Creating admin user"
+      influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -execute="CREATE USER ${ADMIN_USER} WITH PASSWORD '${PASS}' WITH ALL PRIVILEGES"
+      echo "=> Creating database: ${x}"
+      influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} -username=${ADMIN_USER} -password="${PASS}" < /tmp/init_script.influxql
+      echo ""
+    else
+      echo "=> Creating database: ${x}"
+      influx -host=${INFLUX_HOST} -port=${INFLUX_API_PORT} < /tmp/init_script.influxql
+    fi
+
+
+    touch "/data/.init_script_executed"
 else
-    echo "=> No database need to be pre-created"
+    echo "=> No initialization script need to be executed"
 fi
 
 fg
